ci: remove ubuntu-20.04 runner builds on GHA
The Ubuntu 20.04 image on GitHub actions will be unavailable by
2025-04-01, and that is no April Fool's joke. See
<https://github.com/actions/runner-images/issues/11101> for more
information on the deprecation and removal. (Looks like the date
was moved two weeks back to 2025-04-15, but it's time to get rid
of those builds anyway.)

Therefore, the GCC 8 build on Ubuntu 20.04 is removed, and a new
GCC 13 build on Ubuntu 24.04 is added instead. Furthermore, the
Clang 7, 8, 9 and 10 builds on Ubuntu 20.04 are removed, and new
builds for Clang 12 on Ubuntu 22.04 as well as Clang 17, 18 and
19 on Ubuntu 24.04 are added.